requirements.txt
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
python-dotenv==1.0.1
pytest-django==4.10.0